added counter
